+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   * Fixed sub schema unnecessarily outputting when missing
   * Fixed sub schema default value not respecting name key
   * Fixed sub schema mishandling load/name
+  * Fixed parent schema mishandling load/name
 
 
 # 0.6.0